+++ genium/src/main/java/com/here/genium/model/java/JavaMethod.kt
@@ -66,7 +66,7 @@ class JavaMethod(
         returnComment,
         exception,
         parameters,
-        false,
+        isConstructor,
         qualifiers,
         annotations
     )