+++ genium/src/main/java/com/here/genium/model/jni/JniMethod.kt
@@ -26,7 +26,7 @@ class JniMethod(
     val isStatic: Boolean = false,
     val isConst: Boolean = false,
     val isOverloaded: Boolean = false,
-    val isConstructor: Boolean = false,
+    val returnsOpaqueHandle: Boolean = false,
     val exception: JniException? = null,
     val parameters: MutableList<JniParameter> = mutableListOf()
 ) : JniElement