+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   + Java comments are generated as JavaDoc HTML and links to classes and members are supported now
   + Links to members in Swift comments are now supported
   + Added support for structs with methods.
+  + Added support for custom constructors for structs.
 
 ## 4.1.0
 - Release date: 2019-05-06